Making PID determination work on JDK 9;
* Using new Process#getPid() method when running on JDK 9
* Changing type of pid variables to long where needed as Process#getPid() returns a long

+++ src/main/java/de/flapdoodle/embed/process/runtime/ProcessControl.java
@@ -245,8 +245,8 @@ public class ProcessControl {
 	public static void addShutdownHook(Runnable runable) {
 		Runtime.getRuntime().addShutdownHook(new Thread(runable));
 	}
-	
-	public Integer getPid() {
+
+	public Long getPid() {
 		return pid;
 	}
 }
